+++ src/Hacks/NoDeathEffect.cpp
@@ -7,7 +7,14 @@ using namespace geode::prelude;
 class $modify (PlayerObject)
 {
     void playerDestroyed(bool p0)
-    {        
+    {
+        log::info("penis");
+
+        log::info("unk: {}", p0);
+
+        if (Client::GetModuleEnabled("instant-restart") && PlayLayer::get())
+            PlayLayer::get()->resetLevel();
+
         if (!Client::GetModuleEnabled("no-death"))
             PlayerObject::playerDestroyed(p0);
     }