+++ changelog.md
@@ -6,6 +6,7 @@
 - Added No Glow
 - Added Noclip Deaths and Noclip Accuracy
 - Added Alerts to the text bypasses to let the user know that they can crash
+- Added Instant Restart
 - Temporarily removed thicker hitboxes due to lag :(
 
 # 1.1.3